gui: address Copilot review — const wtx in FillWalletTxDetail; explicit int for %n plural
- FillWalletTxDetail takes `const CWalletTx&` (it only reads the tx); the two
  guarded mapValue["from"]/["to"] reads become .at() (the count() checks already
  guarantee presence). FillFinality already took a const ref, so no ripple.
- transactiondesc.cpp: cast finality_value (int64_t) to int for the tr("%n ...")
  plural argument, which is `int` — the OpenForBlocks value is a block count and
  fits int; makes the narrowing explicit.

+++ src/qt/transactiondesc.cpp
@@ -18,7 +18,7 @@ QString TransactionDesc::FormatTxStatus(const WalletTxDetail& detail)
     switch (detail.finality)
     {
     case WalletTxDetail::Finality::OpenForBlocks:
-        return tr("Open for %n more block(s)", "", detail.finality_value);
+        return tr("Open for %n more block(s)", "", static_cast<int>(detail.finality_value));
     case WalletTxDetail::Finality::OpenUntilTime:
         return tr("Open until %1").arg(GUIUtil::dateTimeStr(detail.finality_value));
     case WalletTxDetail::Finality::Conflicted:
